refactor(TemplateLibrary): organize folders - I241
